requirements.txt
beautifulsoup4>=4.6.0
gunicorn>=19.7.1
html5lib
nltk>=3.2.4
passlib>=1.7.1
peewee>=2.10.1
PyMySQL>=0.7.11
python-decouple>=3.0
requests>=2.18.1
tornado>=4.5.1
wtforms-tornado>=0.0.2
